Removed ops leftovers from python.
Removed operation leftovers from python that we forgot to remove
earlier. Also, fixed argmax operation (switch to be reduce op).
