+++ forge/forge/op/eltwise_unary.py
@@ -291,41 +291,6 @@ def Sigmoid(name: str, operandA: Tensor) -> Tensor:
     return op("sigmoid", name, operandA).get_tensor()
 
 
-def Argmax(name: str, operandA: Tensor, dim: int = None, keep_dim=False) -> Tensor:
-    """
-    Argmax
-
-    Parameters
-    ----------
-    name: str
-        Op name, unique to the module, or leave blank to autoset
-
-    operandA: Tensor
-        First operand
-
-    dim: int
-        The dimension to reduce (if None, the output is the argmax of the whole tensor)
-
-    keep_dim: bool
-        If True, retains the dimension that is reduced, with size 1.
-        If False (default), the dimension is removed from the output shape.
-
-    Returns
-    -------
-    Tensor
-        Forge tensor
-    """
-
-    kwargs = {"keep_dim": keep_dim}
-
-    if dim is not None:
-        if dim < 0:
-            dim += len(operandA.shape)
-        kwargs["dim_arg"] = [dim]
-
-    return op("argmax", name, operandA, **kwargs).get_tensor()
-
-
 def Clip(name: str, operandA: Tensor, min: float, max: float) -> Tensor:
     """
     Clips tensor values between min and max
