+++ forge/forge/op/reduce.py
@@ -84,3 +84,38 @@ def ReduceMax(name: str, operandA: Tensor, dim: int, keep_dim: bool = True) -> T
     assert (dim >= -4) and (dim <= 3)
 
     return op("reduce_max", name, operandA, dim_arg=[dim], keep_dim=keep_dim).get_tensor()
+
+
+def Argmax(name: str, operandA: Tensor, dim: int = None, keep_dim=False) -> Tensor:
+    """
+    Argmax
+
+    Parameters
+    ----------
+    name: str
+        Op name, unique to the module, or leave blank to autoset
+
+    operandA: Tensor
+        First operand
+
+    dim: int
+        The dimension to reduce (if None, the output is the argmax of the whole tensor)
+
+    keep_dim: bool
+        If True, retains the dimension that is reduced, with size 1.
+        If False (default), the dimension is removed from the output shape.
+
+    Returns
+    -------
+    Tensor
+        Forge tensor
+    """
+
+    kwargs = {"keep_dim": keep_dim}
+
+    if dim is not None:
+        if dim < 0:
+            dim += len(operandA.shape)
+        kwargs["dim_arg"] = [dim]
+
+    return op("argmax", name, operandA, **kwargs).get_tensor()
